requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
textblob==0.17.1
vadersentiment==3.3.2
matplotlib==3.7.2
seaborn==0.13.0
wordcloud==1.9.3
google-play-scraper==1.2.4
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
plotly==5.18.0
